+++ src/nwbinspector/__init__.py
@@ -2,13 +2,13 @@ from ._version import __version__
 from ._registration import available_checks, register_check
 from ._types import Importance, Severity, InspectorMessage
 from ._configuration import load_config, validate_config, configure_checks
-from ._inspection import (
+from ._nwb_inspection import (
     inspect_all,
     inspect_nwbfile,
     inspect_nwbfile_object,
     run_checks,
 )
-from ._inspection import inspect_nwb  # TODO: remove after 7/1/2023
+from ._nwb_inspection import inspect_nwb  # TODO: remove after 7/1/2023
 from ._formatting import (
     format_messages,
     print_to_console,
@@ -18,6 +18,7 @@ from ._formatting import (
     InspectorOutputJSONEncoder,
 )
 from ._organization import organize_messages
+from ._dandi_inspection import inspect_dandiset, inspect_dandi_file_path, inspect_url
 from .checks import *  # These need to be imported to trigger registration with 'available_checks', but are not exposed
 
 default_check_registry = {check.__name__: check for check in available_checks}
@@ -33,6 +34,9 @@ __all__ = [
     "load_config",
     "configure_checks",
     "InspectorOutputJSONEncoder",
+    "inspect_dandiset",
+    "inspect_dandi_file_path",
+    "inspect_url",
     "inspect_all",
     "inspect_nwbfile",
     "inspect_nwbfile_object",